+++ js/const.js
@@ -1,0 +1,31 @@
+const NAMES = ['Вячеслав', 'Эвелина', 'Анастасия', 'Максим', 'Артём', 'Евгений', 'Елена', 'Екатерина', 'Павел', 'Арсений', 'Эмилия', 'Егор'];
+
+const DESCRIPTION = [
+  'Все только начинает становиться действительно хорошим.',
+  'Сегодня я проснулась вот так ? А вы ?',
+  'Я никогда ничего не теряю. Я либо побеждаю, либо узнаю что-то новое.',
+  'Выходные, пожалуйста, не оставляйте меня сейчас!',
+  'Мне нравится и улыбаюсь!',
+  'Я просто увидел самого умного человека в мире, когда смотрел в зеркало.' ];
+
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!' ];
+
+const AVATAR = {
+  MIN: 1,
+  MAX: 6
+};
+
+const LIKES = {
+  MIN: 15,
+  MAX: 200
+};
+
+const MAX_PHOTO = 25;
+
+export {NAMES, DESCRIPTION, MESSAGE, AVATAR, LIKES, MAX_PHOTO};
