+++ package.json
@@ -11,12 +11,11 @@
     "url": "https://github.com/htmlacademy-univer-javascript-1/2221365-kekstagram-1/issues"
   },
   "devDependencies": {
-    "browser-sync": "2.27.10",
+    "browser-sync": "^2.11.2",
     "eslint": "8.15.0",
     "eslint-config-htmlacademy": "5.1.2"
   },
   "license": "MIT",
-  "dependencies": {},
   "scripts": {
     "start": "browser-sync start --server --no-ui --files \"js/**/*.js\"",
     "lint": "eslint js/"
